moved months and added greying class
so basically, i was trying so hard to find the .hol that i completely forgot i changed it. hol is now bk, which obviously stands for bisha kale. i love coding multilingually! however, this will probably be an issue for anyone else editing since i doubt that'll be so obvious even to another Soomaali person so will be part of my polish plan inshallah